Cleaning the Backend part

+++ Server/Controllers/UserController.ts
@@ -12,7 +12,7 @@ export const getUser = async (req: Request, res: Response): Promise<void> => {
         if (user) {
             res.status(200).json(user);
         } else {
-            res.status(404).json("User Doesn't exist");
+            res.status(404).json({message: "User Doesn't exist"});
         }
     } catch (error) {
         res.status(500).json(error);
@@ -22,7 +22,7 @@ export const getUser = async (req: Request, res: Response): Promise<void> => {
 // Getting all users
 export const getAllUsers = async (req: Request, res: Response) : Promise<void> => {
     try {
-        const users: User[] = await UserModel.find();
+        const users: User[] = await UserModel.find().select('-password').lean<User[]>().exec();
         res.status(200).send(users);
     } catch (error) {
         res.status(500).json(error);
@@ -40,33 +40,36 @@ export const updateUser = async (req: Request, res: Response): Promise<void> => 
                 const salt = await bcrypt.genSalt(10);
                 req.body.password = await bcrypt.hash(password, salt);
             }
-            const user: User | null = await UserModel.findByIdAndUpdate(id, req.body, { new: true });
+            const user: User | null = await UserModel.findByIdAndUpdate(id, req.body, { new: true }).select('-password').lean<User>().exec();;
 
             res.status(200).json(user);
         } else {
-            res.status(403).json("You are not allowed to update this Profile");
+            res.status(403).json({message: "You are not allowed to update this Profile"});
         }
-    } catch (error) {
-        res.status(500).json(error);
+    } catch (error: any) {
+        res.status(500).json({ message: error.message });
     }
 };
 
 // Deleting a User
 export const deleteUser = async (req: Request, res: Response): Promise<void> => {
     const id: string = req.params.id;
+    const { currentUserId, currentUserAdminStatus, password }: { currentUserId: string, currentUserAdminStatus: boolean, password?: string } = req.body;
 
     try {
-
         const user: User | null = await UserModel.findById(id).select('-password').lean<User>().exec();
 
-        if(user) {
-            await UserModel.findByIdAndDelete(id);
-            res.status(200).json("Account deleted successfully");
+        if (user) {
+            if (id === currentUserId || currentUserAdminStatus) {
+                await UserModel.findByIdAndDelete(id);
+                res.status(200).json({ message: "Account deleted successfully" });
+            } else {
+                res.status(403).json({ message: "You are not allowed to delete this account" });
+            }
+        } else {
+            res.status(200).json({ message: "User Account doesn't exist" });
         }
-        else {
-            res.status(200).json("User Account doesn't exist");
-        }
-    } catch (error) {
-        res.status(500).json(error);
+    } catch (error: any) {
+        res.status(500).json({ message: error.message });
     }
 };
